修改部分工具类

+++ app/src/main/java/com/company/project/util/CountObserver.java
@@ -3,7 +3,7 @@ package com.company.project.util;
 import io.reactivex.Observer;
 import io.reactivex.disposables.Disposable;
 
-public abstract class CountObserver implements Observer<Integer> {
+public abstract class CountObserver implements Observer<Long> {
 
     @Override
     public void onSubscribe(Disposable d) {
@@ -11,7 +11,7 @@ public abstract class CountObserver implements Observer<Integer> {
     }
 
     @Override
-    public abstract void onNext(Integer t);
+    public abstract void onNext(Long t);
 
     @Override
     public void onError(Throwable e) {
